Update Assignment6.c
in the above code the program scans input of number of elements of array then elements after that it sorts the array in ascending order .
After sorting the array it evaluates the expression with all elements [in line 34] i.e is (a[current element]!=a[previous element]) true? if yes then count of distinct variable increased by 1.